+++ common/autoware_motion_utils/include/autoware/motion_utils/trajectory_container/trajectory.hpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY_HPP_
-#define AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY_HPP_
-#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_path_point.hpp"
-#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_path_point_with_lane_id.hpp"
-#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_point.hpp"
-#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_pose.hpp"
-#endif  // AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY_HPP_
+#ifndef AUTOWARE__MOTION_UTILS__TRAJECTORY_CONTAINER__TRAJECTORY_HPP_
+#define AUTOWARE__MOTION_UTILS__TRAJECTORY_CONTAINER__TRAJECTORY_HPP_
+#include "autoware/motion_utils/trajectory_container/trajectory/trajectory_path_point.hpp"
+#include "autoware/motion_utils/trajectory_container/trajectory/trajectory_path_point_with_lane_id.hpp"
+#include "autoware/motion_utils/trajectory_container/trajectory/trajectory_point.hpp"
+#include "autoware/motion_utils/trajectory_container/trajectory/trajectory_pose.hpp"
+#endif  // AUTOWARE__MOTION_UTILS__TRAJECTORY_CONTAINER__TRAJECTORY_HPP_
