+++ common/autoware_motion_utils/include/autoware/motion_utils/trajectory_container/trajectory/trajectory_path_point_with_lane_id.hpp
@@ -12,20 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY__TRAJECTORY_PATH_POINT_WITH_LANE_ID_HPP_
-#define AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY__TRAJECTORY_PATH_POINT_WITH_LANE_ID_HPP_
-#include "autoware/motion_utils/trajectory_v2/detail/manipulable_interpolated_array.hpp"
-#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory.hpp"
-#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_path_point.hpp"
+#ifndef AUTOWARE__MOTION_UTILS__TRAJECTORY_CONTAINER__TRAJECTORY__TRAJECTORY_PATH_POINT_WITH_LANE_ID_HPP_  // NOLINT
+#define AUTOWARE__MOTION_UTILS__TRAJECTORY_CONTAINER__TRAJECTORY__TRAJECTORY_PATH_POINT_WITH_LANE_ID_HPP_  // NOLINT
+#include "autoware/motion_utils/trajectory_container/detail/manipulable_interpolated_array.hpp"
+#include "autoware/motion_utils/trajectory_container/trajectory/trajectory.hpp"
+#include "autoware/motion_utils/trajectory_container/trajectory/trajectory_path_point.hpp"
 
 #include "tier4_planning_msgs/msg/path_point_with_lane_id.hpp"
 
 #include <memory>
 #include <vector>
 
-namespace autoware::motion_utils::trajectory_v2::trajectory
+namespace autoware::motion_utils::trajectory_container::trajectory
 {
-using motion_utils::trajectory_v2::detail::ManipulableInterpolatedArray;
+using motion_utils::trajectory_container::detail::ManipulableInterpolatedArray;
 using tier4_planning_msgs::msg::PathPointWithLaneId;
 
 /**
@@ -76,8 +76,8 @@ public:
   std::vector<PathPointWithLaneId> restore() const;
 };
 
-}  // namespace autoware::motion_utils::trajectory_v2::trajectory
+}  // namespace autoware::motion_utils::trajectory_container::trajectory
 
 // clang-format off
-#endif  // AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY__TRAJECTORY_PATH_POINT_WITH_LANE_ID_HPP_  // NOLINT
+#endif  // AUTOWARE__MOTION_UTILS__TRAJECTORY_CONTAINER__TRAJECTORY__TRAJECTORY_PATH_POINT_WITH_LANE_ID_HPP_  // NOLINT
 // clang-format on
