+++ common/autoware_motion_utils/src/trajectory_v2/interpolator/linear.cpp
@@ -12,14 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/motion_utils/trajectory_v2/interpolator/linear.hpp"
+#include "autoware/motion_utils/trajectory_container/interpolator/linear.hpp"
 
 #include <Eigen/Dense>
 
 #include <algorithm>
 #include <vector>
 
-namespace autoware::motion_utils::trajectory_v2::interpolator
+namespace autoware::motion_utils::trajectory_container::interpolator
 {
 
 void Linear::build_(
@@ -60,4 +60,4 @@ size_t Linear::minimum_required_points() const
   return 2;
 }
 
-}  // namespace autoware::motion_utils::trajectory_v2::interpolator
+}  // namespace autoware::motion_utils::trajectory_container::interpolator
