+++ common/autoware_motion_utils/src/trajectory_v2/trajectory/trajectory_path_point.cpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_path_point.hpp"
+#include "autoware/motion_utils/trajectory_container/trajectory/trajectory_path_point.hpp"
 
-#include "autoware/motion_utils/trajectory_v2/detail/merge_vector.hpp"
-#include "autoware/motion_utils/trajectory_v2/interpolator/zero_order_hold.hpp"
+#include "autoware/motion_utils/trajectory_container/detail/merge_vector.hpp"
+#include "autoware/motion_utils/trajectory_container/interpolator/zero_order_hold.hpp"
 
-namespace autoware::motion_utils::trajectory_v2::trajectory
+namespace autoware::motion_utils::trajectory_container::trajectory
 {
 
 TrajectoryV2<PathPoint>::TrajectoryV2() : BaseClass()
@@ -81,7 +81,7 @@ PathPoint TrajectoryV2<PathPoint>::compute(const double & s) const
 
 std::vector<PathPoint> TrajectoryV2<PathPoint>::restore() const
 {
-  auto new_axis = trajectory_v2::detail::merge_vectors(
+  auto new_axis = trajectory_container::detail::merge_vectors(
     axis_, longitudinal_velocity_mps.axis_, lateral_velocity_mps.axis_, heading_rate_rps.axis_);
 
   std::vector<PathPoint> points(new_axis.size());
@@ -97,4 +97,4 @@ std::vector<PathPoint> TrajectoryV2<PathPoint>::restore() const
   return points;
 }
 
-}  // namespace autoware::motion_utils::trajectory_v2::trajectory
+}  // namespace autoware::motion_utils::trajectory_container::trajectory
